+++ icd-web-client/src/main/scala/icd/web/client/IcdWebClient.scala
@@ -14,6 +14,8 @@ import scala.concurrent.ExecutionContext.Implicits.global
 import BrowserHistory._
 import Components._
 import icd.web.client.SelectDialog.SelectDialogListener
+import org.scalajs.dom.ext.Ajax
+import play.api.libs.json.Json
 
 /**
  * Main class for the ICD web app.
@@ -38,7 +40,7 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
   private val sidebar       = Sidebar(LeftSidebarListener)
 
   private val historyItem    = NavbarItem("History", "Display the version history for an API or ICD", showVersionHistory())
-  private val versionHistory = VersionHistory(mainContent)
+  private val historyDialog = HistoryDialog(mainContent)
 
   private val pdfItem = NavbarItem("PDF", "Generate and display a PDF for the API or ICD", makePdf)
 
@@ -51,25 +53,41 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
   private val selectItem   = NavbarItem("Select", "Select the API or ICD to display", selectSubsystems())
   private val selectDialog = SelectDialog(mainContent, Selector)
 
-  private val fileUploadItem   = NavbarItem("Upload", "Select icd model files to ingest into the icd database", uploadSelected())
+  private val fileUploadItem   = NavbarItem("Upload", "Select icd model files to ingest into the icd database", showUploadDialog())
   private val fileUploadDialog = FileUploadDialog(subsystemNames, csrfToken, inputDirSupported)
 
-  // Call popState() when the user presses the browser Back button
-  dom.window.onpopstate = popState _
+  private val publishItem   = NavbarItem("Publish", "Shows dialog to publish APIs and ICDs", showPublishDialog())
+  private val publishDialog = PublishDialog(mainContent, subsystemNames)
 
-  // Initial browser state
-  doLayout()
-  selectSubsystems()
+  isUploadAllowed.map { uploadAllowed =>
+    // Call popState() when the user presses the browser Back button
+    dom.window.onpopstate = popState _
+
+    // Initial browser state
+    doLayout(uploadAllowed)
+    selectSubsystems()
+  }
+
+  // See if uploading model files is allowed in this configuration
+  private def isUploadAllowed: Future[Boolean] = {
+    val path = Routes.isUploadAllowed
+    Ajax.get(path).map { r =>
+      val response = Json.fromJson[Boolean](Json.parse(r.responseText)).get
+      response
+    }
+  }
+
 
   // Layout the components on the page
-  private def doLayout(): Unit = {
+  private def doLayout(uploadAllowed: Boolean): Unit = {
     // Add CSS styles
     head.appendChild(Styles.render[TypedTag[HTMLStyleElement]].render)
 
     navbar.addItem(selectItem)
-    navbar.addItem(fileUploadItem)
+    if (uploadAllowed) navbar.addItem(fileUploadItem)
     navbar.addItem(historyItem)
     navbar.addItem(pdfItem)
+    navbar.addItem(publishItem)
     navbar.addItem(expandToggler)
 
     layout.addItem(sidebar)
@@ -115,10 +133,18 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
   }
 
   // Called when the Upload item is selected
-  private def uploadSelected(saveHistory: Boolean = true)(): Unit = {
+  private def showUploadDialog(saveHistory: Boolean = true)(): Unit = {
     setSidebarVisible(false)
     mainContent.setContent(fileUploadDialog, "Upload Subsystem Model Files")
     if (saveHistory) pushState(viewType = UploadView)
+  }
+
+  // Called when the Publish item is selected
+  private def showPublishDialog(saveHistory: Boolean = true)(): Unit = {
+    setSidebarVisible(false)
+    showBusyCursorWhile(publishDialog.update())
+    mainContent.setContent(publishDialog, "Publish APIs and ICDs")
+    if (saveHistory) pushState(viewType = PublishView)
   }
 
   // Listener for sidebar component checkboxes
@@ -205,12 +231,13 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
       e.preventDefault()
       // Make sure to wait for futures to complete, so things happen in the right order
       for {
-        _ <- selectDialog.subsystem.setSubsystemWithVersion(hist.maybeSourceSubsystem, saveHistory = false)
-        _ <- selectDialog.targetSubsystem.setSubsystemWithVersion(hist.maybeTargetSubsystem, saveHistory = false)
+        _ <- selectDialog.subsystem.setSubsystemWithVersion(hist.maybeSourceSubsystem, saveHistory = false, findMatchingIcd = false)
+        _ <- selectDialog.targetSubsystem.setSubsystemWithVersion(hist.maybeTargetSubsystem, saveHistory = false, findMatchingIcd = false)
         _ <- selectDialog.icdChooser.setIcdWithVersion(hist.maybeIcd, notifyListener = false, saveHistory = false)
       } {
         hist.viewType match {
-          case UploadView  => uploadSelected(saveHistory = false)()
+          case UploadView  => showUploadDialog(saveHistory = false)()
+          case PublishView  => showPublishDialog(saveHistory = false)()
           case VersionView => showVersionHistory(saveHistory = false)()
           case SelectView =>
             selectSubsystems(
@@ -232,21 +259,6 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
         }
       }
     }
-  }
-
-  // Show/hide the busy cursor while the future is running
-  private def showBusyCursorWhile(f: Future[Unit]): Future[Unit] = {
-    // Note: See implicit NodeList to List support in package object in this dir
-    val nodeList = document.querySelectorAll("div")
-    nodeList.map(_.asInstanceOf[HTMLDivElement]).foreach { divEl =>
-      divEl.style.cursor = "progress"
-    }
-    f.onComplete { _ =>
-      nodeList.map(_.asInstanceOf[HTMLDivElement]).foreach { divEl =>
-        divEl.style.cursor = "default"
-      }
-    }
-    f
   }
 
   private object Selector extends SelectDialogListener {
@@ -294,16 +306,16 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
   private def showVersionHistory(saveHistory: Boolean = true)(): Unit = {
     selectDialog.icdChooser.getSelectedIcd match {
       case Some(icdName) =>
-        versionHistory.setIcd(icdName)
+        historyDialog.setIcd(icdName)
         setSidebarVisible(false)
-        mainContent.setContent(versionHistory, s"ICD Version History: ${icdName.subsystem} to ${icdName.target}")
+        mainContent.setContent(historyDialog, s"ICD Version History: ${icdName.subsystem} to ${icdName.target}")
         if (saveHistory) pushState(viewType = VersionView)
       case None =>
         selectDialog.subsystem.getSelectedSubsystem match {
           case Some(name) =>
-            versionHistory.setSubsystem(name)
+            historyDialog.setSubsystem(name)
             setSidebarVisible(false)
-            mainContent.setContent(versionHistory, s"Subsystem API Version History: $name")
+            mainContent.setContent(historyDialog, s"Subsystem API Version History: $name")
             if (saveHistory) pushState(viewType = VersionView)
           case None =>
         }