+++ icd-web-shared/src/main/scala/icd/web/shared/JsonSupport.scala
@@ -69,5 +69,9 @@ object JsonSupport {
   implicit val diffInfoFormat              = Json.format[DiffInfo]
   implicit val versionInfoFormat           = Json.format[VersionInfo]
   implicit val icdVersionInfoFormat        = Json.format[IcdVersionInfo]
+  implicit val apiVersionInfoFormat        = Json.format[ApiVersionInfo]
+  implicit val publishInfoFormat           = Json.format[PublishInfo]
+  implicit val publishApiInfoFormat        = Json.format[PublishApiInfo]
+  implicit val publishIcdInfoFormat        = Json.format[PublishIcdInfo]
 
 }