+++ icd-web-shared/src/main/scala/icd/web/shared/SubsystemInfo.scala
@@ -1,7 +1,5 @@
 package icd.web.shared
 
-object SubsystemInfo {}
-
 /**
  * Holds information about a subsystem
  *